+++ src/backend/marsha/core/api/timed_text_track.py
@@ -1,6 +1,5 @@
 """Declare API endpoints for videos with Django RestFramework viewsets."""
 from django.conf import settings
-from django.utils import timezone
 
 import django_filters
 from rest_framework import filters, viewsets
@@ -8,7 +7,7 @@ from rest_framework.decorators import action
 from rest_framework.exceptions import MethodNotAllowed
 from rest_framework.response import Response
 
-from marsha.core import defaults, permissions, serializers
+from marsha.core import defaults, permissions, serializers, storage
 from marsha.core.api.base import (
     APIViewMixin,
     ObjectPkMixin,
@@ -17,8 +16,7 @@ from marsha.core.api.base import (
 )
 from marsha.core.metadata import TimedTextMetadata
 from marsha.core.models import TimedTextTrack
-from marsha.core.utils.s3_utils import create_presigned_post
-from marsha.core.utils.time_utils import to_timestamp
+from marsha.core.tasks.timed_text_track import convert_timed_text_track
 
 
 class TimedTextTrackFilter(django_filters.FilterSet):
@@ -69,6 +67,7 @@ class TimedTextTrackViewSet(
             "retrieve",
             "update",
             "partial_update",
+            "upload_ended",
         ]:
             permission_classes = [
                 permissions.IsTokenPlaylistRouteObjectRelatedVideo
@@ -117,18 +116,51 @@ class TimedTextTrackViewSet(
         serializer = serializers.TimedTextUploadSerializer(data=request.data)
         serializer.is_valid(raise_exception=True)
 
-        now = timezone.now()
-        stamp = to_timestamp(now)
-
-        key = timed_text_track.get_source_s3_key(stamp=stamp)
-
-        presigned_post = create_presigned_post(
-            [["content-length-range", 0, settings.SUBTITLE_SOURCE_MAX_SIZE]],
-            {},
-            key,
+        presigned_post = (
+            storage.get_initiate_backend().initiate_object_videos_storage_upload(
+                request,
+                timed_text_track,
+                [["content-length-range", 0, settings.SUBTITLE_SOURCE_MAX_SIZE]],
+            )
         )
 
         # Reset the upload state of the timed text track
         TimedTextTrack.objects.filter(pk=pk).update(upload_state=defaults.PENDING)
 
         return Response(presigned_post)
+
+    @action(methods=["post"], detail=True, url_path="upload-ended")
+    # pylint: disable=unused-argument
+    def upload_ended(self, request, pk=None, video_id=None):
+        """Notify the API that the timed text track upload has ended.
+
+        Calling the endpoint will start the conversion of the timed text track.
+        The request should have a file_key in the body, which is the key of the
+        uploaded file.
+
+        Parameters
+        ----------
+        request : Type[django.http.request.HttpRequest]
+            The request on the API endpoint
+        pk: string
+            The primary key of the timed text track
+
+        Returns
+        -------
+        Type[rest_framework.response.Response]
+            HttpResponse with the serialized timed text track.
+        """
+        # Ensure object exists and user has access to it
+        timed_text_track = self.get_object()
+
+        serializer = serializers.VideosStorageUploadEndedSerializer(
+            data=request.data, context={"obj": timed_text_track}
+        )
+        serializer.is_valid(raise_exception=True)
+
+        file_key = serializer.validated_data["file_key"]
+        # The file_key have the "tmp/{video_pk}/thumbnail/{stamp}" format
+        stamp = file_key.split("/")[-1]
+        convert_timed_text_track.delay(pk, stamp)
+
+        return Response(serializer.data)