+++ src/backend/marsha/core/models/video.py
@@ -594,6 +594,14 @@ class TimedTextTrack(BaseTrack):
         verbose_name=_("extension"),
     )
 
+    process_pipeline = models.CharField(
+        max_length=255,
+        verbose_name=_("process pipeline"),
+        help_text=_("Pipeline used to process the timed text track"),
+        choices=PROCESS_PIPELINE_CHOICES,
+        default=CELERY_PIPELINE,
+    )
+
     class Meta:
         """Options for the ``TimedTextTrack`` model."""
 
@@ -633,6 +641,37 @@ class TimedTextTrack(BaseTrack):
         stamp = stamp or to_timestamp(self.uploaded_on)
         mode = f"_{self.mode}" if self.mode else ""
         return f"{self.video.pk}/timedtexttrack/{self.pk}/{stamp}_{self.language}{mode}"
+
+    def get_videos_storage_prefix(
+        self,
+        stamp=None,
+        base_dir: VIDEOS_STORAGE_BASE_DIRECTORY = VOD_VIDEOS_STORAGE_BASE_DIRECTORY,
+    ):
+        """Compute the videos storage prefix for the video.
+
+        Parameters
+        ----------
+        stamp: Type[string]
+            Passing a value for this argument will return the videos storage prefix for the video
+            assuming its active stamp is set to this value. This is useful to create an upload
+            policy for this prospective version of the video, so that the client can upload the
+            file to S3 and the transcodings job can set the `uploaded_on` field to this value.
+
+        base_dir: Type[VIDEOS_STORAGE_BASE_DIRECTORY]
+            The videos storage base directory. Defaults to VOD. It will be used to compute the
+            videos storage prefix.
+
+        Returns
+        -------
+        string
+            The videos storage prefix for the video, depending on the base directory passed.
+        """
+        stamp = stamp or to_timestamp(self.uploaded_on)
+        base = base_dir
+        if base_dir == DELETED_VIDEOS_STORAGE_BASE_DIRECTORY:
+            base = f"{DELETED_VIDEOS_STORAGE_BASE_DIRECTORY}/{VOD_VIDEOS_STORAGE_BASE_DIRECTORY}"
+
+        return f"{base}/{self.video.pk}/timedtext/{self.pk}/{stamp}"
 
     def update_upload_state(self, upload_state, uploaded_on, **extra_parameters):
         """Manage upload state.